+++ xknx/io/tunnel.py
@@ -189,10 +189,7 @@ class _Tunnel(Interface):
             self.communication_channel = connect.communication_channel
             # assign data_endpoint received from server
             self._data_endpoint_addr = (
-                (
-                    connect.data_endpoint.ip_addr,
-                    connect.data_endpoint.port,
-                )
+                connect.data_endpoint.addr_tuple
                 if not connect.data_endpoint.route_back
                 else None
             )